move product.js
